Enable the ros2 integration tests and clean them up
